Separate header main content from bottom content
