+++ h5p-bildetema/src/components/Header/Header.module.scss
@@ -3,6 +3,9 @@
 .header {
   background-color: $burgundy;
   padding: $spacing--32 0;
+  position: sticky;
+  top: 0;
+  z-index: 2;
 
   &_content {
     align-items: center;
@@ -63,26 +66,6 @@
   flex-wrap: wrap;
 }
 
-.bottom {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  height: 2.5rem;
-  margin: $main-content-margin;
-  max-width: $main-content-max-width;
-  width: $main-content-width;
-}
-
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: $spacing--32;
-}
-
-.toggle {
-  font-size: 1.2rem;
-}
-
 .languageButton,
 .languageMenuButton {
   background-color: transparent;
@@ -97,6 +80,7 @@
     text-decoration: underline;
   }
 
+  &_active,
   &.active {
     text-decoration: underline;
     &:focus {
